Update beta script

+++ scripts/beta-release.js
@@ -38,6 +38,8 @@ fs.writeFileSync(path.join(__dirname, '../README.md'), readme);
 
 // Update the .vscodeignore file
 const ignoreFilePath = path.join(path.resolve('.'), '.vscodeignore');
-let vscodeignore = fs.readFileSync(ignoreFilePath, 'utf8');
-vscodeignore = vscodeignore.replace(`**/*.map`, '');
-fs.writeFileSync(ignoreFilePath, vscodeignore);
+if (fs.existsSync(ignoreFilePath)) {
+  let vscodeignore = fs.readFileSync(ignoreFilePath, 'utf8');
+  vscodeignore = vscodeignore.replace(`**/*.map`, '');
+  fs.writeFileSync(ignoreFilePath, vscodeignore);
+}
